backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
tensorflow-cpu
scikit-learn
plotly
pandas-ta
python-binance
joblib
